python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "my-assist"
version = "0.1.0"
description = "Local-first LLM assistant with script-to-tool promotion"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "My Assist Team" }
]

dependencies = [
    "microsoft-teams-ai>=2.0.0a8",
    "microsoft-teams-apps>=2.0.0a8",
    "microsoft-teams-devtools>=2.0.0a8",
    "pydantic-ai>=0.1.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "ulid-py>=1.1.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
my-assist = "assistant.main:main"
agcom-api = "agcom_api.main:run"
agcom = "agcom.console.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["assistant", "agcom", "agcom_api"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
